Adds local storage to save Policies data.

+++ app/src/main/java/me/wooz/mobile/android/app/policies/AddPolicyActivity.java
@@ -224,11 +224,11 @@ public class AddPolicyActivity extends BaseActivity {
 	 */
 	public class AddPolicyTask extends AsyncTask<Void, Void, Boolean> {
 
-		private Integer insuranceCompanyId;
-		private Integer insuranceTypeId;
+		private Long insuranceCompanyId;
+		private Long insuranceTypeId;
 		private String policyNumber;
 
-		public AddPolicyTask(Integer insuranceCompanyId, Integer insuranceTypeId,
+		public AddPolicyTask(Long insuranceCompanyId, Long insuranceTypeId,
 				String policyNumber) {
 			this.insuranceCompanyId = insuranceCompanyId;
 			this.insuranceTypeId = insuranceTypeId;
